Add AC API specific test client.
