+++ aiarena/api/arenaclient/tests.py
@@ -10,6 +10,7 @@ from aiarena.core.models import Match, Bot, MatchParticipation, User, Round, Res
     ArenaClient, competition_participation
 from aiarena.core.models.bot_race import BotRace
 from aiarena.core.models.game_mode import GameMode
+from aiarena.core.tests.testing_utils import TestAssetPaths
 from aiarena.core.tests.tests import LoggedInMixin, MatchReadyMixin
 from aiarena.core.utils import calculate_md5
 from aiarena.settings import ELO_START_VALUE, BASE_DIR, PRIVATE_STORAGE_ROOT, MEDIA_ROOT
@@ -22,10 +23,11 @@ class MatchesTestCase(LoggedInMixin, TransactionTestCase):
                                                      email='regular_user2@aiarena.net')
 
     def test_get_next_match_not_authorized(self):
+        self.test_ac_api_client.set_api_token(None)
         response = self._post_to_matches()
         self.assertEqual(response.status_code, 403)
 
-        self.client.login(username='regular_user', password='x')
+        self.test_ac_api_client.set_api_token(Token.objects.get(user=self.regularUser1).key)
         response = self._post_to_matches()
         self.assertEqual(response.status_code, 403)
 
@@ -34,7 +36,6 @@ class MatchesTestCase(LoggedInMixin, TransactionTestCase):
         config.REISSUE_UNFINISHED_MATCHES = False
 
         self.test_client.login(self.staffUser1)
-        self.test_client.set_api_token(Token.objects.get(user=self.arenaclientUser1))
 
         # no current competition
         response = self._post_to_matches()
@@ -44,7 +45,6 @@ class MatchesTestCase(LoggedInMixin, TransactionTestCase):
         comp = self._create_game_mode_and_open_competition()
 
         # no maps
-        self.test_client.login(self.arenaclientUser1)
         response = self._post_to_matches()
         self.assertEqual(response.status_code, 200)
         self.assertEqual(u'no_current_competitions', response.data['detail'].code)
@@ -127,7 +127,7 @@ class MatchesTestCase(LoggedInMixin, TransactionTestCase):
         self._create_active_bot_for_competition(comp.id, self.regularUser1, 'testbot1', BotRace.terran())
         self._create_active_bot_for_competition(comp.id, self.regularUser1, 'testbot2', BotRace.zerg())
 
-        self.test_client.login(self.arenaclientUser1)
+        # self.test_client.login(self.arenaclientUser1)
         response_m1 = self.client.post('/api/arenaclient/matches/')
         self.assertEqual(response_m1.status_code, 201)
 
@@ -152,7 +152,7 @@ class MatchesTestCase(LoggedInMixin, TransactionTestCase):
         bot1 = self._create_bot(self.regularUser1, 'testbot1', BotRace.terran())
         bot2 = self._create_bot(self.regularUser1, 'testbot2', BotRace.zerg())
 
-        self.test_client.login(self.arenaclientUser1)
+        # self.test_client.login(self.arenaclientUser1)
 
         # we shouldn't be able to get a new match
         response = self._post_to_matches()
@@ -183,7 +183,7 @@ class MatchesTestCase(LoggedInMixin, TransactionTestCase):
         bot4 = self._create_active_bot_for_competition(comp.id, self.regularUser1, 'testbot4', BotRace.random())
 
         # Round 1
-        self.test_client.login(self.arenaclientUser1)
+        # self.test_client.login(self.arenaclientUser1)
         response = self._post_to_matches()
         self.assertEqual(response.status_code, 201)
         response = self._post_to_matches()
@@ -220,7 +220,7 @@ class MatchesTestCase(LoggedInMixin, TransactionTestCase):
         bot1 = self._create_active_bot_for_competition(comp.id,  self.regularUser1, 'testbot1', BotRace.terran())
         bot2 = self._create_active_bot_for_competition(comp.id, self.regularUser1, 'testbot2', BotRace.zerg())
 
-        self.test_client.login(self.arenaclientUser1)
+        # self.test_client.login(self.arenaclientUser1)
         # this should tie up both bots
         response = self.client.post('/api/arenaclient/matches/')
         self.assertEqual(response.status_code, 201)
@@ -248,7 +248,6 @@ class ResultsTestCase(LoggedInMixin, TransactionTestCase):
 
     def test_create_results(self):
         self.test_client.login(self.staffUser1)
-        self.test_client.set_api_token(Token.objects.get(user=self.arenaclientUser1))
 
         comp = self._create_game_mode_and_open_competition()
         self._create_map_for_competition('test_map', comp.id)
@@ -342,15 +341,14 @@ class ResultsTestCase(LoggedInMixin, TransactionTestCase):
         match_bot = MatchParticipation.objects.get(bot=bot1,
                                                    match_id=match_id)  # use this to determine which hash to match
         if match_bot.participant_number == 1:
-            self.assertEqual(self.test_bot_datas['bot1'][data_index]['hash'], Bot.objects.get(id=bot1.id).bot_data_md5hash)
-            self.assertEqual(self.test_bot_datas['bot2'][data_index]['hash'], Bot.objects.get(id=bot2.id).bot_data_md5hash)
+            self.assertEqual(TestAssetPaths.test_bot_datas['bot1'][data_index]['hash'], Bot.objects.get(id=bot1.id).bot_data_md5hash)
+            self.assertEqual(TestAssetPaths.test_bot_datas['bot2'][data_index]['hash'], Bot.objects.get(id=bot2.id).bot_data_md5hash)
         else:
-            self.assertEqual(self.test_bot_datas['bot1'][data_index]['hash'], Bot.objects.get(id=bot2.id).bot_data_md5hash)
-            self.assertEqual(self.test_bot_datas['bot2'][data_index]['hash'], Bot.objects.get(id=bot1.id).bot_data_md5hash)
+            self.assertEqual(TestAssetPaths.test_bot_datas['bot1'][data_index]['hash'], Bot.objects.get(id=bot2.id).bot_data_md5hash)
+            self.assertEqual(TestAssetPaths.test_bot_datas['bot2'][data_index]['hash'], Bot.objects.get(id=bot1.id).bot_data_md5hash)
 
     def test_create_result_bot_not_in_match(self):
         self.test_client.login(self.staffUser1)
-        self.test_client.set_api_token(Token.objects.get(user=self.arenaclientUser1))
 
         comp = self._create_game_mode_and_open_competition()
         self._create_map_for_competition('test_map', comp.id)
@@ -391,8 +389,6 @@ class ResultsTestCase(LoggedInMixin, TransactionTestCase):
 
         bot1 = self._create_active_bot_for_competition(comp.id, self.regularUser1, 'bot1')
         bot2 = self._create_active_bot_for_competition(comp.id, self.regularUser1, 'bot2', BotRace.zerg())
-
-        self.test_client.set_api_token(Token.objects.get(user=self.arenaclientUser1))
 
         # log more crashes than should be allowed
         for count in range(config.BOT_CONSECUTIVE_CRASH_LIMIT):
@@ -463,7 +459,7 @@ class EloTestCase(LoggedInMixin, TransactionTestCase):
         self.regularUserBot1 = self._create_bot(self.regularUser1, 'regularUserBot1')
         self.regularUserBot2 = self._create_bot(self.regularUser1, 'regularUserBot2')
 
-        self.test_client.login(self.arenaclientUser1)
+        # self.test_client.login(self.arenaclientUser1)
 
         # activate the required bots
         self.regularUserBot1.competition_participations.create(competition=comp)
